made command have help
