+++ Package.swift
@@ -12,6 +12,7 @@ let package = Package(
     dependencies: [
         .package(url: "https://github.com/swift-embedded/swift-package-manager", .branch("embedded-5.1")),
         .package(url: "https://github.com/dduan/TOMLDecoder", from: "0.1.3"),
+        .package(url: "https://github.com/apple/swift-argument-parser", from: "0.0.1"),
     ],
     targets: [
         .target(
@@ -20,7 +21,7 @@ let package = Package(
         ),
         .target(
             name: "cross",
-            dependencies: ["CrossLib"]
+            dependencies: ["CrossLib", .product(name: "ArgumentParser", package: "swift-argument-parser"),]
         ),
     ]
 )
